app/requirements/base.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
sqlparse==0.4.2
psycopg2==2.9.3
pandas==1.4.3
djangorestframework==3.13.1
django-redis==5.2.0
